Issue #927356 by omaster: Enforce content locks when reverting nodes to prior revisions.
Now, the page where the user is asked to confirm performing a revision
causes a lock to be grabbed on the node. This means that no user can
revert a node when another user has it locked for editing. It also
means that noone can edit a node when another is about to revert it.